Update PublishedData OpenAPI descriptions
Motivation:

Many of the fields exposed by the PublishedData endpoint currently have
rather limited descriptions, with poorly defined semantics.

Modification:

Semantics of some fields were understood by examining how they are
currently used both when minting a DOI and when building OAI-PMH
records.  The OpenAPI descriptions are updated following this
understanding, by including the corresponding terms and including
MarkDown links that point to the definition.

(Note: the links point to DataCite metadata v4.5 descriptions because it
is with this version that the schema documentation was available as
HTML.)

An example DOI is added to the description of the `doi` field, to remove
any ambiguity on how the DOI is represented.

The description of the `pidArray` field is updated to remove ambiguity
resulting from "pid" (meaning "persistent identifier") being an
overloaded term.

Result:

The PublishedData endpoint semantics are now better described.

Closes: #1175