Fixup optimistic concurrency control
If the time send by the client is equal to the ressource on the server,
the clients updated version is based on servers version with that time.

or in other words:

the client has taken into account the ressource on the server b/c the
client should use the timestamp received with the ressource from the
server.

from the RFC:

> The origin server MUST NOT perform the requested method
> if the selected representation's last modification date is more
> recent than the date provided in the field-value

Client receives t0
Client sends update w/ t0 -> OK

Server is on t1
Client received t0 (previously)
Client sends update w/ t0 -> FAIL

Change-Id: Ibdff18c17a8b43523cec88ac8b9c16a94fd77617